requirements.txt
pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.10.0
pandas>=2.0.0
numpy>=1.24.0

